Removed anything related with routeNodes.
Also the first commit I tried to enlist some values with an actionlistener, this does not work.
