--- src/main/java/nl/tudelft/pixelperfect/route/AsteroidNode.java
@@ -1,33 +1,0 @@
-package nl.tudelft.pixelperfect.route;
-
-/**
- * An asteroid field Node which belongs in a Route. The asteroid field will be a dangerous location
- * to fly in.
- * 
- * @author Floris Doolaard
- *
- */
-public class AsteroidNode extends RouteNode {
-
-  /**
-   * The constructor for the AsteroidNode.
-   * 
-   * @param summary
-   *          A summary of the AsteroidNode.
-   * @param description
-   *          A description of the AsteroidNode.
-   */
-  public AsteroidNode(String summary, String description) {
-    super(summary, description);
-  }
-
-  /**
-   * String representation of an AsteroidNode.
-   * 
-   * @return A String.
-   */
-  public String toString() {
-    String res = "AsteroidNode";
-    return res;
-  }
-}
