+++ src/test/java/nl/tudelft/pixelperfect/SpaceshipTest.java
@@ -19,9 +19,9 @@ import com.jme3.network.HostedConnection;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import nl.tudelft.pixelperfect.event.EventListener;
 import nl.tudelft.pixelperfect.event.EventLog;
+import nl.tudelft.pixelperfect.game.Route;
 import nl.tudelft.pixelperfect.game.Spaceship;
 import nl.tudelft.pixelperfect.player.Player;
-import nl.tudelft.pixelperfect.route.Route;
 
 /**
  * Test Suite for the Spaceship class.
